% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centros_df.R
\docType{data}
\name{centros_df}
\alias{centros_df}
\title{Datos de centros meteorológicos}
\format{
Un dataframe con \code{n} filas y \code{m} columnas:
\describe{
\item{nombre_centro}{(factor) Nombre del centro meteorológico.}
\item{latitud}{(numérico) Latitud del centro en grados.}
\item{longitud}{(numérico) Longitud del centro en grados.}
\item{altitud}{(numérico) Altitud del centro en metros.}
\item{fecha_inicio}{(Fecha) Fecha de inicio de operaciones del centro.}
}
}
\source{
Datos ficticios para demostración.
}
\usage{
data(centros_df)
}
\description{
Este conjunto de datos contiene información sobre centros meteorológicos, incluyendo
su nombre, ubicación geográfica y fecha de inicio de operaciones.
}
\keyword{datasets}
