% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentros.R
\name{getCentros}
\alias{getCentros}
\title{Descarga de centros}
\usage{
getCentros(centros = NULL, destfile)
}
\arguments{
\item{centros}{Un vector donde se indica los centros que se quieren descargar}

\item{destfile}{Ruta donde se descargarán los .csv}
}
\value{
Si la descarga fue exitosa devuelve una lista de datasets
}
\description{
Función que permite descargar en un directorio los centros que se desean
}
\examples{
getCentros(c("NH0910", "NH0472"), file.path(tempdir()) )

}
\seealso{
\link{centros_df} , \code{\link[=getNH0472]{getNH0472()}}, \code{\link[=getNH0437]{getNH0437()}}, \code{\link[=getNH0098]{getNH0098()}}, \code{\link[=getNH0046]{getNH0046()}}
}
\author{
JM y VM
}
