% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0098.R
\name{getNH0098}
\alias{getNH0098}
\title{Descargar NH0098}
\usage{
getNH0098(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0098
}
\examples{
getNH0098(file.path(tempdir(), "NH0098.csv"))
}
\seealso{
\code{\link[=getNH0910]{getNH0910()}}, \code{\link[=getNH0472]{getNH0472()}}, \code{\link[=getNH0437]{getNH0437()}}, \code{\link[=getNH0046]{getNH0046()}}
}
\author{
JM y VM
}
