% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0437.R
\name{getNH0437}
\alias{getNH0437}
\title{Descargar NH0437}
\usage{
getNH0437(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0437
}
\examples{
getNH0437("datos/NH0437.csv")
}
\author{
JM y VM
}
