% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0472.R
\name{getNH0472}
\alias{getNH0472}
\title{Descargar NH0472}
\usage{
getNH0472(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0472
}
\examples{
getNH0472("NH0472.csv")
}
\seealso{
\code{\link[=getNH0910]{getNH0910()}}, \code{\link[=getNH0437]{getNH0437()}}, \code{\link[=getNH0098]{getNH0098()}}, \code{\link[=getNH0046]{getNH0046()}}
}
\author{
JM y VM
}
