% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0910.R
\name{getNH0910}
\alias{getNH0910}
\title{Descargar NH0910}
\usage{
getNH0910(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0910
}
\examples{
getNH0910("NH0910.csv")
}
\author{
JM y VM
}
