% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatos.R
\docType{data}
\name{metadatos}
\alias{metadatos}
\title{Metadatos de centros metereologicos}
\format{
Un dataframe con \code{258} filas y \code{11} columnas:
\describe{
\item{id}{(chr) Id del centro meteorológico.}
\item{nombre}{(chr) Nombre del centro}
\item{tipo}{(chr) Tipo del centro}
\item{localidad}{(chr) Localidad del centro}
\item{provincia}{(chr) Provincia donde se encuentra el centro}
\item{lat}{(dbl) Latitud geografica del centro}
\item{lon}{(dbl) Longitud geografica del centro}
\item{altura}{(dbl) Altura del nivel del mar del centro}
\item{ubicacion}{(chr) Direccion del centro}
}
}
\source{
"https://github.com/rse-r/intro-programacion/blob/main/datos/metadatos_completos.csv"
}
\usage{
data(metadatos)
}
\description{
Este conjunto de datos contiene metadatos de los centros metereologicos
}
\seealso{
\link{centros_df}
}
\keyword{datasets}
