% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_centro_csv.R
\name{read_centro_csv}
\alias{read_centro_csv}
\title{Leer CSV de un centro desde URL}
\usage{
read_centro_csv(centro)
}
\arguments{
\item{centro}{(str) Nombre del centro a descargar}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
La funcion importa a Rstudio el dataset del centro que queramos desde github, \strong{SIN DESCARGAR EL ARCHIVO}
}
\examples{
NH0910 <- read_centro_csv("NH0910")
NH0910
}
\seealso{
\code{\link[=getNH0910]{getNH0910()}}, \code{\link[=getNH0472]{getNH0472()}}, \code{\link[=getNH0437]{getNH0437()}}, \code{\link[=getNH0098]{getNH0098()}}, \code{\link[=getNH0046]{getNH0046()}}
}
\author{
JM y VM
}
