% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_Estaciones.R
\name{Summary_Temperatura_Abrigo}
\alias{Summary_Temperatura_Abrigo}
\title{Summary de temperatura abrigo}
\usage{
Summary_Temperatura_Abrigo(...)
}
\arguments{
\item{...}{Centros metereologicos}
}
\value{
tibble
}
\description{
La funcion genera una summary para mostrar los estadisticos de los centros que solicites
}
\examples{
centros <- NimbusR::centros_df
Summary_Temperatura_Abrigo(centros)
}
\author{
JM y VM
}
