% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempCtoF.R
\name{TempCtoF}
\alias{TempCtoF}
\title{Temperature Celsius To Farenheit}
\usage{
TempCtoF(temperature)
}
\arguments{
\item{temperature}{Temperatura en Celsius (numeric)}
}
\value{
Temperatura en Farenheit (numeric)
}
\description{
Convierte la temperatura de Celsius a Farenheit
}
\examples{
TempCtoF(10)
}
\author{
JM y VM
}
