% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempFtoC.R
\name{TempFtoC}
\alias{TempFtoC}
\title{Temperature Farenheit To Celsius}
\usage{
TempFtoC(temperature)
}
\arguments{
\item{temperature}{Temperatura en Farenheit (numeric)}
}
\value{
Temperatura en Celsius (numeric)
}
\description{
Convierte la temperatura de Farenheit  a Celsius
}
\examples{
TempFtoc(32)
}
\author{
JM y VM
}
