% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centros_df.R
\docType{data}
\name{centros_df}
\alias{centros_df}
\title{Datos de centros meteorológicos}
\format{
Un dataframe con \verb{95,681} filas y \code{35} columnas:
\describe{
\item{id}{(chr) Nombre del centro meteorológico.}
\item{fecha}{(date) Fecha de la observación.}
\item{temperatura_abrigo_150cm}{(dbl) Temperatura abrigo media.}
\item{temperatura_abrigo_150cm_maxima}{(dbl) Temperatura abrigo máxima}
\item{temperatura_abrigo_150cm_minima}{(dbl) Temperatura abrigo mínima.}
}
}
\source{
Datos ficticios para demostración.
}
\usage{
data(centros_df)
}
\description{
Este conjunto de datos contiene información sobre los centros meteorológicos NH0910, NH0472, NH0437,
NH0098 y NH0046
}
\keyword{datasets}
