% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0046.R
\name{getNH0046}
\alias{getNH0046}
\title{Descargar NH0046}
\usage{
getNH0046(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0046
}
\examples{
getNH0046("NH0046.csv")
}
\author{
JM y VM
}
