% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNH0437.R
\name{getNH0437}
\alias{getNH0437}
\title{Descargar NH0437}
\usage{
getNH0437(destfile)
}
\arguments{
\item{destfile}{Ruta donde se descargará el .csv}
}
\value{
Si la descarga fue exitosa devuelve el dataset
}
\description{
Descarga los datos del centro NH0437
}
\examples{
getNH0437("NH0437.csv")
}
\seealso{
\code{\link[=getNH0910]{getNH0910()}}, \code{\link[=getNH0472]{getNH0472()}}, \code{\link[=getNH0098]{getNH0098()}}, \code{\link[=getNH0046]{getNH0046()}}
}
\author{
JM y VM
}
