% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperatura_abrigo.R
\name{plot_temperatura_abrigo}
\alias{plot_temperatura_abrigo}
\title{Serie de tiempo temperatura abrigo}
\usage{
plot_temperatura_abrigo(dataset)
}
\arguments{
\item{dataset}{Dataset con datos de los centros (Deben tener columna "temperatura_abrigo_150cm")}
}
\value{
Gráfico de la temperatura abrigo.
}
\description{
Genera un grafico de serie de tiempo que muestra la temperatura abrigo de cada estacion
}
\author{
JM y VM
}
