% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{Suma}
\alias{Suma}
\title{Suma de numeros}
\usage{
Suma(x = 0, y = 0)
}
\value{
Devuelve la suma de dos numeros positivos
}
\description{
Suma valores positivos.
}
\examples{
Suma(x = 5 , y = 2)
}
\author{
JM y VM
}
